requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==19.3.0
Automat==0.8.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.1
blinker==1.4
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
executing==0.10.0
fastjsonschema==2.16.1
Flask==2.2.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.12.0
incremental==16.10.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
more-itertools==4.2.0
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.12
numpy==1.23.2
oauthlib==3.1.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.1.1
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-docx==0.8.11
pytz==2022.2.1
PyYAML==5.3.1
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.2.0
sos==4.3
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.4.0
systemd-python==234
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
Twisted==18.9.0
typing-extensions==4.3.0
ubuntu-advantage-tools==27.10
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.2
zipp==1.0.0
zope.interface==4.7.1
